requirements.txt
Flask==0.12
requests
kafka-python==2.0.2